# contracts/admin-api.openapi.yaml
openapi: 3.1.0
info:
  title: AI Control Layer – Admin API
  description: Management and reporting API consumed by the admin panel.
  version: 1.0.0
  x-gateway-version: 0.1.0
paths:
...
  /admin/v1/policy:
    get:
      tags:
      - policy
      summary: Get Status
      operationId: getPolicyStatus
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyStatus'
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
      security:
      - HTTPBearer: []
  /admin/v1/policy/schema:
    get:
      tags:
      - policy
      summary: Get Schema
      description: JSON Schema for policy files (same as contracts/policy.schema.json).
      operationId: getPolicySchema
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Getpolicyschema
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
      security:
      - HTTPBearer: []
  /admin/v1/policy/files:
    get:
      tags:
      - policy
      summary: List Files
      operationId: listPolicyFiles
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/PolicyFileInfo'
                type: array
                title: Response Listpolicyfiles
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
      security:
      - HTTPBearer: []
  /admin/v1/policy/files/{name}:
    get:
      tags:
      - policy
      summary: Get File
      operationId: getPolicyFile
      security:
      - HTTPBearer: []
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
          title: Name
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyFileContent'
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - policy
      summary: Write File
      description: 'Single-writer update: validate → round-trip write (comments preserved) → reload → snapshot.'
      operationId: writePolicyFile
      security:
      - HTTPBearer: []
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
          title: Name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyFileWrite'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyStatus'
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '409':
          description: Stale base_version
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation failed or locked control edited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/v1/policy/validate:
    post:
      tags:
      - policy
      summary: Validate
      operationId: validatePolicy
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidateResponse'
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /admin/v1/policy/dry-run:
    post:
      tags:
      - policy
      summary: Dry Run
      description: Replay the last N stored requests against the candidate policy.
      operationId: dryRunPolicy
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DryRunRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DryRunResponse'
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /admin/v1/policy/versions:
    get:
      tags:
      - policy
      summary: List Versions
      operationId: listPolicyVersions
      security:
      - HTTPBearer: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PolicyVersion'
                title: Response Listpolicyversions
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/v1/policy/versions/{version_id}:
    get:
      tags:
      - policy
      summary: Get Version
      operationId: getPolicyVersion
      security:
      - HTTPBearer: []
      parameters:
      - name: version_id
        in: path
        required: true
        schema:
          type: integer
          title: Version Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyVersionDetail'
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/v1/policy/versions/{version_id}/rollback:
    post:
      tags:
      - policy
      summary: Rollback
      operationId: rollbackPolicy
      security:
      - HTTPBearer: []
      parameters:
      - name: version_id
        in: path
        required: true
        schema:
          type: integer
          title: Version Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyStatus'
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    Action:
      type: string
      enum:
      - allow
      - monitor
      - redact
      - pseudonymise
      - sanitize
      - route_local
      - downgrade
      - require_approval
      - block
      title: Action
      description: Decision actions (concept §6.3).
...
    DryRunChange:
      properties:
        event_id:
          type: string
          title: Event Id
        timestamp:
          type: string
          format: date-time
          title: Timestamp
        subject:
          anyOf:
          - type: string
          - type: 'null'
          title: Subject
        before_action:
          $ref: '#/components/schemas/Action'
        after_action:
          $ref: '#/components/schemas/Action'
        before_rule_ids:
          items:
            type: string
            pattern: ^[A-Z][A-Z0-9]*(-[A-Z0-9_.]+)+$
            description: Rule identifier (control id, or control id + sub-rule, or signature id).
          type: array
          title: Before Rule Ids
        after_rule_ids:
          items:
            type: string
            pattern: ^[A-Z][A-Z0-9]*(-[A-Z0-9_.]+)+$
            description: Rule identifier (control id, or control id + sub-rule, or signature id).
          type: array
          title: After Rule Ids
      additionalProperties: false
      type: object
      required:
      - event_id
      - timestamp
      - before_action
      - after_action
      title: DryRunChange
    DryRunRequest:
      properties:
        files:
          additionalProperties:
            type: string
          type: object
          title: Files
          description: Candidate file contents by name; omitted files use the current version.
        last_n:
          type: integer
          maximum: 10000.0
          minimum: 1.0
          title: Last N
          default: 500
        point:
          anyOf:
          - $ref: '#/components/schemas/InspectionPoint'
          - type: 'null'
      additionalProperties: false
      type: object
      required:
      - files
      title: DryRunRequest
    DryRunResponse:
      properties:
        candidate_version:
          type: string
          title: Candidate Version
        evaluated:
          type: integer
          title: Evaluated
        changed:
          type: integer
          title: Changed
        transitions:
          additionalProperties:
            type: integer
          type: object
          title: Transitions
          description: '`allow->block` → count.'
        samples:
          items:
            $ref: '#/components/schemas/DryRunChange'
          type: array
          title: Samples
        errors:
          items:
            $ref: '#/components/schemas/PolicyError'
          type: array
          title: Errors
      additionalProperties: false
      type: object
      required:
      - candidate_version
      - evaluated
      - changed
      title: DryRunResponse
...
    ErrorResponse:
      properties:
        error:
          type: string
          title: Error
          examples:
          - stale_version
          - validation_failed
          - forbidden_model
          - not_implemented
        message:
          type: string
          title: Message
        details:
          additionalProperties: true
          type: object
          title: Details
      additionalProperties: false
      type: object
      required:
      - error
      - message
      title: ErrorResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    InspectionPoint:
      type: string
      enum:
      - ingress
      - egress
      - tool_call
      - tool_result
      - embeddings
      - agent_message
      - artifact_load
      - mcp_initialize
      - mcp_tools_list
      title: InspectionPoint
      description: Where in the traffic flow a payload is inspected (concept §6.1).
...
    PolicyError:
      properties:
        file:
          anyOf:
          - type: string
          - type: 'null'
          title: File
        path:
          anyOf:
          - type: string
          - type: 'null'
          title: Path
          description: Location inside the document, e.g. `groups.developers.preset`.
        line:
          anyOf:
          - type: integer
          - type: 'null'
          title: Line
        column:
          anyOf:
          - type: integer
          - type: 'null'
          title: Column
        message:
          type: string
          title: Message
      additionalProperties: false
      type: object
      required:
      - message
      title: PolicyError
    PolicyFileContent:
      properties:
        name:
          type: string
          title: Name
          examples:
          - controls.yaml
        version:
          type: string
          title: Version
          description: Content hash; use as `base_version` for optimistic locking.
        size:
          type: integer
          title: Size
        modified_at:
          type: string
          format: date-time
          title: Modified At
        content:
          type: string
          title: Content
      additionalProperties: false
      type: object
      required:
      - name
      - version
      - size
      - modified_at
      - content
      title: PolicyFileContent
    PolicyFileInfo:
      properties:
        name:
          type: string
          title: Name
          examples:
          - controls.yaml
        version:
          type: string
          title: Version
          description: Content hash; use as `base_version` for optimistic locking.
        size:
          type: integer
          title: Size
        modified_at:
          type: string
          format: date-time
          title: Modified At
      additionalProperties: false
      type: object
      required:
      - name
      - version
      - size
      - modified_at
      title: PolicyFileInfo
    PolicyFileWrite:
      properties:
        content:
          type: string
          title: Content
        base_version:
          type: string
          title: Base Version
          description: Version the edit was based on; stale → 409.
        message:
          type: string
          maxLength: 500
          title: Message
          default: ''
      additionalProperties: false
      type: object
      required:
      - content
      - base_version
      title: PolicyFileWrite
    PolicyStatus:
      properties:
        version:
          type: string
          title: Version
        loaded_at:
          type: string
          format: date-time
          title: Loaded At
        source:
          type: string
          enum:
          - file
          - panel
          - rollback
          - startup
          title: Source
        files:
          items:
            $ref: '#/components/schemas/PolicyFileInfo'
          type: array
          title: Files
        last_error:
          items:
            $ref: '#/components/schemas/PolicyError'
          type: array
          title: Last Error
          description: Errors from the last failed reload.
        locked_controls:
          items:
            type: string
          type: array
          title: Locked Controls
      additionalProperties: false
      type: object
      required:
      - version
      - loaded_at
      - source
      - files
      title: PolicyStatus
    PolicyVersion:
      properties:
        id:
          type: integer
          title: Id
        version:
          type: string
          title: Version
        created_at:
          type: string
          format: date-time
          title: Created At
        author:
          anyOf:
          - type: string
          - type: 'null'
          title: Author
        source:
          type: string
          enum:
          - file
          - panel
          - rollback
          - startup
          title: Source
        message:
          type: string
          title: Message
          default: ''
        files_changed:
          items:
            type: string
          type: array
          title: Files Changed
      additionalProperties: false
      type: object
      required:
      - id
      - version
      - created_at
      - source
      title: PolicyVersion
    PolicyVersionDetail:
      properties:
        id:
          type: integer
          title: Id
        version:
          type: string
          title: Version
        created_at:
          type: string
          format: date-time
          title: Created At
        author:
          anyOf:
          - type: string
          - type: 'null'
          title: Author
        source:
          type: string
          enum:
          - file
          - panel
          - rollback
          - startup
          title: Source
        message:
          type: string
          title: Message
          default: ''
        files_changed:
          items:
            type: string
          type: array
          title: Files Changed
        diff:
          type: string
          title: Diff
          description: Unified diff against the previous version.
        files:
          additionalProperties:
            type: string
          type: object
          title: Files
      additionalProperties: false
      type: object
      required:
      - id
      - version
      - created_at
      - source
      - diff
      title: PolicyVersionDetail
...
    ValidateRequest:
      properties:
        files:
          additionalProperties:
            type: string
          type: object
          title: Files
          description: Candidate file contents by name; omitted files use the current version.
      additionalProperties: false
      type: object
      required:
      - files
      title: ValidateRequest
    ValidateResponse:
      properties:
        valid:
          type: boolean
          title: Valid
        errors:
          items:
            $ref: '#/components/schemas/PolicyError'
          type: array
          title: Errors
        candidate_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Candidate Version
      additionalProperties: false
      type: object
      required:
      - valid
      title: ValidateResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      description: Keycloak access token (JWT) or personal API key (`acl_...`).
      scheme: bearer
servers:
- url: http://localhost:8000